Removed Commons project.
Moved regex and validation constants to Data project. Moved validation error messages to Service project.
