+++ Development/Data/Validation/Commons/CustomValidationRules.cs
@@ -1,4 +1,4 @@
-﻿using Commons;
+﻿using Data.Messages;
 using FluentValidation;
 using System;
 using System.Text.RegularExpressions;
