+++ Development/Service/Service.csproj
@@ -9,7 +9,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Commons\Commons.csproj" />
     <ProjectReference Include="..\Data\Data.csproj" />
   </ItemGroup>
