+++ Development/Service/Service/DictionaryService.cs
@@ -2,7 +2,7 @@
 using Data.Repository;
 using System.Collections.Generic;
 using System.Linq;
-using Msg = Commons.ValidationErrorMessages;
+using Msg = Service.ValidationErrorMessages;
 
 namespace Service
 {
